move macros into utils.rs
