+++ rust/candid/src/utils.rs
@@ -1,6 +1,41 @@
 use crate::de::IDLDeserialize;
+use crate::ser::IDLBuilder;
 use crate::{CandidType, Result};
 use serde::de::Deserialize;
+
+/// Encode sequence of Rust values into Candid message of type `candid::Result<Vec<u8>>`.
+#[macro_export]
+macro_rules! Encode {
+    ( $($x:expr),* ) => {{
+        let mut builder = $crate::ser::IDLBuilder::new();
+        Encode!(@PutValue builder $($x,)*)
+    }};
+    ( @PutValue $builder:ident $x:expr, $($tail:expr,)* ) => {{
+        $builder.arg($x).and_then(|builder| Encode!(@PutValue builder $($tail,)*))
+    }};
+    ( @PutValue $builder:ident ) => {{
+        $builder.serialize_to_vec()
+    }};
+}
+
+/// Decode Candid message into a tuple of Rust values of the given types.
+/// Produces `Err` if the message fails to decode at any given types.
+/// If the message contains only one value, it returns the value directly instead of a tuple.
+#[macro_export]
+macro_rules! Decode {
+    ( $hex:expr $(,$ty:ty)* ) => {{
+        $crate::de::IDLDeserialize::new($hex)
+            .and_then(|mut de| Decode!(@GetValue [] de $($ty,)*)
+                      .and_then(|res| de.done().and(Ok(res))))
+    }};
+    (@GetValue [$($ans:ident)*] $de:ident $ty:ty, $($tail:ty,)* ) => {{
+        $de.get_value::<$ty>()
+            .and_then(|val| Decode!(@GetValue [$($ans)* val] $de $($tail,)* ))
+    }};
+    (@GetValue [$($ans:ident)*] $de:ident) => {{
+        Ok(($($ans),*))
+    }};
+}
 
 /// Decode a series of arguments, represented as a tuple. There is a maximum of 16 arguments
 /// supported.
@@ -49,6 +84,63 @@ where
     Ok(res)
 }
 
+/// Serialize an encoding of a tuple and write it to a [Write] buffer.
+///
+/// ```
+/// # use candid::Decode;
+/// # use candid::ser::write_args;
+/// let golden1 = 1u64;
+/// let golden2 = "hello";
+/// let mut buffer = Vec::new();
+/// write_args(&mut buffer, (golden1, golden2)).unwrap();
+///
+/// let (value1, value2) = Decode!(&buffer, u64, String).unwrap();
+/// assert_eq!(golden1, value1);
+/// assert_eq!(golden2, value2);
+/// ```
+pub fn write_args<Tuple: ArgumentEncoder, Writer: std::io::Write>(
+    writer: &mut Writer,
+    arguments: Tuple,
+) -> Result<()> {
+    let mut ser = IDLBuilder::new();
+    arguments.encode(&mut ser)?;
+    ser.serialize(writer)
+}
+
+/// Serialize an encoding of a tuple to a vector of bytes.
+///
+/// ```
+/// # use candid::Decode;
+/// # use candid::ser::encode_args;
+/// let golden1 = 1u64;
+/// let golden2 = "hello";
+/// let buffer = encode_args((golden1, golden2)).unwrap();
+///
+/// let (value1, value2) = Decode!(&buffer, u64, String).unwrap();
+/// assert_eq!(golden1, value1);
+/// assert_eq!(golden2, value2);
+/// ```
+pub fn encode_args<Tuple: ArgumentEncoder>(arguments: Tuple) -> Result<Vec<u8>> {
+    let mut result = Vec::new();
+    write_args(&mut result, arguments)?;
+    Ok(result)
+}
+
+/// Serialize a single value to a vector of bytes.
+///
+/// ```
+/// # use candid::Decode;
+/// # use candid::ser::encode_one;
+/// let golden = "hello";
+/// let buffer = encode_one(golden).unwrap();
+///
+/// let (value) = Decode!(&buffer, String).unwrap();
+/// assert_eq!(golden, value);
+/// ```
+pub fn encode_one<T: CandidType>(argument: T) -> Result<Vec<u8>> {
+    encode_args((argument,))
+}
+
 /// Allow decoding of any sized argument.
 pub trait ArgumentDecoder<'a>: Sized {
     /// Decodes a value of type [Self], modifying the deserializer (values are consumed).
@@ -58,6 +150,19 @@ pub trait ArgumentDecoder<'a>: Sized {
 /// Decode an empty tuple.
 impl<'a> ArgumentDecoder<'a> for () {
     fn decode(_de: &mut IDLDeserialize<'a>) -> Result<()> {
+        Ok(())
+    }
+}
+
+/// Allow encoding of any serializable value.
+pub trait ArgumentEncoder {
+    /// Encode a value of type [Self].
+    fn encode(self, ser: &mut IDLBuilder) -> Result<()>;
+}
+
+/// Decode an empty tuple.
+impl ArgumentEncoder for () {
+    fn encode(self, _de: &mut IDLBuilder) -> Result<()> {
         Ok(())
     }
 }
@@ -75,6 +180,25 @@ macro_rules! decode_impl {
                 )*
 
                 Ok(($( $id, )*))
+            }
+        }
+    }
+}
+
+// Create implementation of [ArgumentEncoder] for up to 16 value tuples.
+macro_rules! encode_impl {
+    ( $($id: ident : $typename: ident),* ) => {
+        impl<$( $typename ),*> ArgumentEncoder for ($($typename,)*)
+        where
+            $( $typename: CandidType ),*
+        {
+            fn encode(self, ser: &mut IDLBuilder) -> Result<()> {
+                let ( $( $id, )* ) = self;
+                $(
+                ser.arg(&$id)?;
+                )*
+
+                Ok(())
             }
         }
     }
@@ -102,3 +226,26 @@ decode_impl!(a: A, b: B, c: C, d: D, e: E, f: F, g: G, h: H, i: I, j: J, k: K, l
 decode_impl!(a: A, b: B, c: C, d: D, e: E, f: F, g: G, h: H, i: I, j: J, k: K, l: L, m: M, n: N, o: O);
 #[rustfmt::skip]
 decode_impl!(a: A, b: B, c: C, d: D, e: E, f: F, g: G, h: H, i: I, j: J, k: K, l: L, m: M, n: N, o: O, p: P);
+
+encode_impl!(a: A);
+encode_impl!(a: A, b: B);
+encode_impl!(a: A, b: B, c: C);
+encode_impl!(a: A, b: B, c: C, d: D);
+encode_impl!(a: A, b: B, c: C, d: D, e: E);
+encode_impl!(a: A, b: B, c: C, d: D, e: E, f: F);
+encode_impl!(a: A, b: B, c: C, d: D, e: E, f: F, g: G);
+encode_impl!(a: A, b: B, c: C, d: D, e: E, f: F, g: G, h: H);
+encode_impl!(a: A, b: B, c: C, d: D, e: E, f: F, g: G, h: H, i: I);
+encode_impl!(a: A, b: B, c: C, d: D, e: E, f: F, g: G, h: H, i: I, j: J);
+#[rustfmt::skip]
+encode_impl!(a: A, b: B, c: C, d: D, e: E, f: F, g: G, h: H, i: I, j: J, k: K);
+#[rustfmt::skip]
+encode_impl!(a: A, b: B, c: C, d: D, e: E, f: F, g: G, h: H, i: I, j: J, k: K, l: L);
+#[rustfmt::skip]
+encode_impl!(a: A, b: B, c: C, d: D, e: E, f: F, g: G, h: H, i: I, j: J, k: K, l: L, m: M);
+#[rustfmt::skip]
+encode_impl!(a: A, b: B, c: C, d: D, e: E, f: F, g: G, h: H, i: I, j: J, k: K, l: L, m: M, n: N);
+#[rustfmt::skip]
+encode_impl!(a: A, b: B, c: C, d: D, e: E, f: F, g: G, h: H, i: I, j: J, k: K, l: L, m: M, n: N, o: O);
+#[rustfmt::skip]
+encode_impl!(a: A, b: B, c: C, d: D, e: E, f: F, g: G, h: H, i: I, j: J, k: K, l: L, m: M, n: N, o: O, p: P);
